Rename hooks.yaml -> .pre-commit-hooks.yaml
In https://github.com/pre-commit/pre-commit/pull/470, hooks.yaml was
renamed to .pre-commit-hooks.yaml to try to make the file more likely
for upstreams to include.

New versions of pre-commit will support both, but issue a warning about
the hooks.yaml name.

Moving it and setting up a symlink for compatibility should work on all
versions.